counting formal release (#366)
Move features that have had time to be battle tested in the experimental
section into the main client without change other than minor doc updates

---------
