+++ test/unit/test_experimental.py
@@ -90,51 +90,6 @@ def test_submit_multiple_rois(gl_experimental: ExperimentalApi, image_query_one:
     gl_experimental.add_label(image_query_one, 3, [roi] * 3)
 
 
-def test_counting_detector(gl_experimental: ExperimentalApi):
-    """
-    verify that we can create and submit to a counting detector
-    """
-    name = f"Test {datetime.utcnow()}"
-    created_detector = gl_experimental.create_counting_detector(name, "How many dogs", "dog", confidence_threshold=0.0)
-    assert created_detector is not None
-    count_iq = gl_experimental.submit_image_query(created_detector, "test/assets/dog.jpeg")
-    assert count_iq.result.count is not None
-
-
-def test_counting_detector_async(gl_experimental: ExperimentalApi):
-    """
-    verify that we can create and submit to a counting detector
-    """
-    name = f"Test {datetime.utcnow()}"
-    created_detector = gl_experimental.create_counting_detector(name, "How many dogs", "dog", confidence_threshold=0.0)
-    assert created_detector is not None
-    async_iq = gl_experimental.ask_async(created_detector, "test/assets/dog.jpeg")
-    # attempting to access fields within the result should raise an exception
-    with pytest.raises(AttributeError):
-        _ = async_iq.result.label  # type: ignore
-    with pytest.raises(AttributeError):
-        _ = async_iq.result.confidence  # type: ignore
-    time.sleep(5)
-    # you should be able to get a "real" result by retrieving an updated image query object from the server
-    _image_query = gl_experimental.get_image_query(id=async_iq.id)
-    assert _image_query.result is not None
-
-
-def test_multiclass_detector(gl_experimental: ExperimentalApi):
-    """
-    verify that we can create and submit to a multi-class detector
-    """
-    name = f"Test {datetime.utcnow()}"
-    class_names = ["Golden Retriever", "Labrador Retriever", "Poodle"]
-    created_detector = gl_experimental.create_multiclass_detector(
-        name, "What kind of dog is this?", class_names=class_names, confidence_threshold=0.0
-    )
-    assert created_detector is not None
-    mc_iq = gl_experimental.submit_image_query(created_detector, "test/assets/dog.jpeg")
-    assert mc_iq.result.label is not None
-    assert mc_iq.result.label in class_names
-
-
 def test_text_recognition_detector(gl_experimental: ExperimentalApi):
     """
     verify that we can create and submit to a text recognition detector
